refactor: theme typescript for autocomplete
