+++ packages/@borealisgroup/theme/src/index.ts
@@ -1,0 +1,7 @@
+export {
+  borealisTheme,
+  ecoplastTheme,
+  latTheme,
+  mtmPlasticsTheme,
+  rosierTheme,
+} from './themes';
